Add Require, PubSub, modules
I really believe in modularity. The absolute ideal would be to have
components of Castle Fortify that can be used in completely different
projects, making them more robust, and worth perfecting.

For this to happen, we've got to use RequireJs, which allows us to
pull in dependencies in a standard way.. Kinda like IOC?

Then by using PubSub, a mediator that you can publish and subscribe to,
we can achieve decoupling of the components.

I've seperated the code into data, utilities, and the map_model.
So the dom is now a bit seperate from the map model, and we're no
longer saving data directly from the textarea.

It looks like the map model is emerging as something that benefits from
being decoupled from castle fortify. Everything else is just slightly
easier to read.

Big plus in having the mapToDoctrine formatters reorganized.

Just hope I didn't break anything! I'm well aware of the issues with
this kind of release. Zero new features, just a "better" code base.
Great learning experience though, and I think I'll have an easier
time debugging.